Improve eslint

+++ database/migrations/2019_02_20_060404_create_links_table.php
@@ -27,7 +27,7 @@ class CreateLinksTable extends Migration
             $table->string('hostname')->nullable();
 
             $table->string('hash')->nullable();
-            $table->tinyInteger('clicks')->default(0);
+            $table->integer('clicks')->default(0);
 
             $table->timestamps();
             $table->softDeletes();
